requirements.txt
altgraph==0.17
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
docopt==0.6.2
easygui==0.98.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
macholib==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
pdfminer.six==20200517
pycodestyle==2.6.0
pycryptodome==3.9.8
PyInstaller==3.6
pylint==2.5.3
PyMuPDF==1.17.2
pynsist==2.5.1
PyPDF2==1.26.0
requests==2.24.0
requests-download==0.1.2
six==1.15.0
sortedcontainers==2.2.2
tika==1.24
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
